refactor(useFocus): refactor repeated useEffect code
